Add support for cart context in MarketCart
This commit adds support for cart context in MarketCart, allowing the user to manage multiple items in the cart, including updating and removing them. The changes also include updating ControlersContext to track quantity and properly updating the cart context on any changes. Additionally, a new CartContext file was created to support the new functionality of keeping track of multiple items in the cart.