Add pot number to OrderNow component
This change adds a new function to the OrderNow component that calculates the number of cactus pots in the cart and displays it above the total price. The function loops through the items in the cart and sums up their quantities. Now users can easily see how many pots they're ordering at a glance.