Add Framer Motion package & new pages and images
This commit adds the `framer-motion` package, which provides animation support to the project. It also includes new pages, `Contact` and `About`, as well as additional images for pots and cactuses. Lastly, a new import statement is added to include the new images. Some components have also been updated to include animation from `framer-motion`.